refactor: split dead code result checkbox
